refactor(fbuild-build): split esp32/orchestrator.rs into submodules to satisfy LOC gate
